requirements.txt
Flask==1.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==1.2.13
alembic==1.0.2
flask-jwt-extended==3.13.1
Flask-Whooshee
python-dotenv==0.9.1
flask-restful==0.3.6
flask-marshmallow==0.9.0
passlib==1.7.1
marshmallow-sqlalchemy==0.15.0
Flask-Cors==3.0.7
psycopg2==2.7.6.1
marshmallow==2.16.3